analytics/proof_of_correlation/requirements.txt
# HRKey Correlation Engine - Python Dependencies
# Phase I: Proof of Correlation (MVP)

# Core data science libraries
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Database
psycopg2-binary>=2.9.0

# Configuration
python-dotenv>=1.0.0

# API (optional - for serving results)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Type checking (optional)
mypy>=1.5.0

# Code quality (optional)
black>=23.9.0
flake8>=6.1.0
isort>=5.12.0
